crawling/requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.2
Django==4.0.5
django-ninja==0.19.0
h11==0.13.0
happybase==1.2.0
idna==3.3
kafka-python==2.0.2
oauthlib==3.2.0
opensearch-py==2.0.0
outcome==1.2.0
ply==3.11
pycparser==2.21
pydantic==1.9.1
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
requests==2.28.1
requests-oauthlib==1.3.1
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
thriftpy2==0.4.14
trio==0.21.0
trio-websocket==0.9.2
tweepy==3.10.0
typing-extensions==4.3.0
urllib3==1.26.9
webdriver-manager==3.7.1
wsproto==1.1.0
